+++ handlers/MapLibreMapController.MaciOS.cs
@@ -227,6 +227,8 @@ public class MapLibreMapController : IMapLibreMapController
     public void SetCompassViewMargins(int x, int y)           { }
     public void SetAttributionButtonGravity(int v)            { }
     public void SetAttributionButtonMargins(int x, int y)     { }
+    public void SetShowNavigationControls(bool show)          { }
+    public void SetShowAttributionControl(bool show, string? customAttribution) { }
 
     // -- Sources ---------------------------------------------------------------