+++ handlers/MapLibreMapHandler.Android.cs
@@ -59,4 +59,6 @@ public partial class MapLibreMapHandler : ViewHandler<MapLibreMap, SurfaceView>
         if (margin?.Length >= 2 && margin[0] != null && margin[1] != null)
             _controller.SetAttributionButtonMargins(margin[0]!.Value, margin[1]!.Value);
     }
+    public void UpdateShowNavigationControls(bool show) => _controller.SetShowNavigationControls(show);
+    public void UpdateShowAttributionControl(bool show, string? customAttribution) => _controller.SetShowAttributionControl(show, customAttribution);
 }