+++ handlers/MapLibreMapHandler.MaciOS.cs
@@ -46,4 +46,6 @@ public partial class MapLibreMapHandler : ViewHandler<MapLibreMap, UIView>
     public void UpdateCompassViewMargins(int?[]? margin){ if (margin?.Length >= 2 && margin[0] != null && margin[1] != null) _controller.SetCompassViewMargins(margin[0]!.Value, margin[1]!.Value); }
     public void UpdateAttributionButtonGravity(int v)   => _controller.SetAttributionButtonGravity(v);
     public void UpdateAttributionButtonMargins(int?[]? margin) { if (margin?.Length >= 2 && margin[0] != null && margin[1] != null) _controller.SetAttributionButtonMargins(margin[0]!.Value, margin[1]!.Value); }
+    public void UpdateShowNavigationControls(bool show) => _controller.SetShowNavigationControls(show);
+    public void UpdateShowAttributionControl(bool show, string? customAttribution) => _controller.SetShowAttributionControl(show, customAttribution);
 }